% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdcm.R
\name{ggdcm}
\alias{ggdcm}
\title{Quickly plot a DCM in ggplot2}
\usage{
ggdcm(.tdy_dag, int.vars = NULL, lat.vars = NULL, use_labels = FALSE, ...)
}
\arguments{
\item{.tdy_dag}{input graph, an object of class \code{tidy_dagitty} or
\code{dagitty}}

\item{int.vars}{variables names for indicators}

\item{lat.vars}{variables names for latent class or attributes}

\item{use_labels}{a string. Variable to use for \code{geom_dag_repel_label()}.
Default is \code{NULL}.}

\item{...}{additional arguments passed to \code{tidy_dagitty()}}
}
\value{
a \code{ggplot}
}
\description{
\code{ggdcm()} is a wrapper to quickly plot DAGs.
}
\examples{

library(tidyverse)
dag.dcm  <- dagify(
  x1 ~ Attr1 + Attr2,
  x2 ~ Attr1,
  x3 ~ Attr1 + Attr2,
  x4 ~ Attr2 + Attr1,
  x5 ~ Attr2,
  x6 ~ Attr2 + Attr1,
  x7 ~ Attr2,
  labels = c(Attr1 = "Multiplication", Attr2 = "Minus")
)

ggdcm(.tdy_dag = dag.dcm, use_labels = "label", int.vars = paste0("x", 1:7))
ggdcm(.tdy_dag = dag.dcm, use_labels = "label")
ggdcm(.tdy_dag = dag.dcm, int.vars = paste0("x", 1:6), lat.vars = paste0("Attr", 1:2), use_labels = "label")



}
